Use doubles to avoid class error in bxsfun
The first array into bxsfun is int32 class, which throws the error
"mixed integer class inputs are not supported". To avoid this, we wrap
the first array in double(). The second array is already a double.

+++ @iss/call_spots_pixel.m
@@ -102,7 +102,7 @@ for f = 1:nFiles
     load(cell2mat(o.PixelFileNames(f)));
     PeakGlobalYX = cell(nCodes,1);
     for GeneNo = 1:nCodes
-        PeakGlobalYX{GeneNo} = bsxfun(@plus,PeakLocalYX{GeneNo},o.TileOrigin(OriginalTile{GeneNo},:,rr));
+        PeakGlobalYX{GeneNo} = bsxfun(@plus,double(PeakLocalYX{GeneNo}),o.TileOrigin(OriginalTile{GeneNo},:,rr));
     end 
     % Clear memory by removing bad matches, pSpotScore<-5
     clearvars PeakLocalYX
